Make check_packages.py infer the name of the target branch
